Major Refactor so that the cache-image and puppeteer-report don't have to rely on opening the data file again on the server side
Major Refactor so that the cache-image and puppeteer-report don't have to rely on opening the data file again on the server side, instead the client app sends the file to the server, using multer to receive the file as post data. This means the app now works as 'upload the data once only'.